refactor: update to headshot sent by David Giard
